Fix bug that resulted in the order state not being changed
The order state should be changed on the return callback after the
customer has entered all their details and confirmed their afterpay
payment. The registerPaymentReviewAction method sets the order state and
status to processing, but this is not saved in the database. Immediately
after this is done, the sendNewOrderEmail method is called, which loads
the order from the database, resetting the order status back to
pending_payment.

We resolve this issue by saving the order after calling
registerPaymentReviewAction, the same way as it is done in the core
PayPal IPN model.