Fix upgrade scripts and fix bug with order cancellation
The upgrade scripts that were included for different version to 0.13.0
would all be run if someone upgraded from that version. Only one upgrade
script is required to the 0.13.0 version so the rest were removed.

The 0.13.0 install script was incorrect, and didn't include most
changes. If someone fresh installed version 0.13.0 the extension
wouldn't work.

When an order is cancelled the cancel action would find the last order
placed by that customer and cancel it. With the new API method no order
is created when the cancel action is run, so it would find any orders
placed previously. If a customer placed and order that was in the
pending state (eg: a direct deposit method) and then cancelled an
afterpay payment, their original order would be cancelled. This change
adds a check for the API version before continuing with the order
cancellation code.