Bump to version 1.0.0 and update the README
The readme was accidentally reverted to include some typos when the
extension was restructured. An incorrect version of the README was
located at src/main/php/README.md and it was copied into the git root of
the project when the restructure was done. This commit reverts that
accidental change, and also tags 1.0.0 so the package can be added to
packagist. The existing tags are not compatible with composer (0.13.0j
etc).