Update README.md
Copy edit to how to configure Afterpay (1.2) after it's installed. Line 25: Clarifies where to find the option to select/deselect which countries require a postal code.

I used these instructions to install Afterpay on a client's dev site today and I think this clarification will be clearer for future users.